feat: peer-lite
